Refactor pages into components and default to dashboard
