+++ src/app/login/page.tsx
@@ -1,0 +1,5 @@
+import { Login } from '@/components/gear-guardian/login';
+
+export default function LoginPage() {
+  return <Login />;
+}
