+++ src/components/gear-guardian/login.tsx
@@ -1,0 +1,99 @@
+'use client';
+
+import * as React from 'react';
+import Link from 'next/link';
+import { useRouter } from 'next/navigation';
+import { login } from '@/lib/api';
+import { Button } from '@/components/ui/button';
+import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card';
+import { Input } from '@/components/ui/input';
+import { Label } from '@/components/ui/label';
+import { Logo } from '@/components/gear-guardian/logo';
+import { useToast } from '@/hooks/use-toast';
+import { Loader2 } from 'lucide-react';
+
+// Formulaire de connexion affiché sur la page Login
+export function Login() {
+  const router = useRouter();
+  const { toast } = useToast();
+  // Indique si l'appel API est en cours
+  const [isLoading, setIsLoading] = React.useState(false);
+
+  // Gestion de la soumission du formulaire
+  const handleLogin = async (event: React.FormEvent<HTMLFormElement>) => {
+    event.preventDefault();
+    setIsLoading(true);
+    const formData = new FormData(event.currentTarget);
+    const email = formData.get('email') as string;
+    const password = formData.get('password') as string;
+
+    if (!email || !password) {
+      toast({
+        variant: 'destructive',
+        title: 'Champs manquants',
+        description: 'Veuillez renseigner votre email et mot de passe.',
+      });
+      setIsLoading(false);
+      return;
+    }
+
+    try {
+      await login(email, password);
+      router.push('/dashboard');
+    } catch (error: any) {
+      console.error(error);
+      toast({
+        variant: 'destructive',
+        title: 'Erreur de connexion',
+        description: "L'email ou le mot de passe est incorrect.",
+      });
+      setIsLoading(false);
+    }
+  };
+
+  return (
+    <div className="flex items-center justify-center min-h-screen bg-muted/40">
+      <Card className="w-full max-w-sm mx-4">
+        <CardHeader className="text-center">
+          <div className="flex justify-center items-center gap-2 mb-4">
+            <Logo className="size-8" />
+            <h1 className="text-2xl font-headline">GearGuardian</h1>
+          </div>
+          <CardTitle className="text-2xl">Connexion</CardTitle>
+          <CardDescription>Entrez votre email ci-dessous pour vous connecter à votre compte.</CardDescription>
+        </CardHeader>
+        <CardContent>
+          <form onSubmit={handleLogin}>
+            <div className="grid gap-4">
+              <div className="grid gap-2">
+                <Label htmlFor="email">Email</Label>
+                <Input id="email" name="email" type="email" placeholder="nom@example.com" required />
+              </div>
+              <div className="grid gap-2">
+                <div className="flex items-center">
+                  <Label htmlFor="password">Mot de passe</Label>
+                  <Link href="#" className="ml-auto inline-block text-sm underline">
+                    Mot de passe oublié?
+                  </Link>
+                </div>
+                <Input id="password" name="password" type="password" required />
+              </div>
+              <Button type="submit" className="w-full" disabled={isLoading}>
+                {isLoading && <Loader2 className="mr-2 h-4 w-4 animate-spin" />}
+                Se connecter
+              </Button>
+            </div>
+          </form>
+          <div className="mt-4 text-center text-sm">
+            Vous n'avez pas de compte?{' '}
+            <Link href="/signup" className="underline">
+              S'inscrire
+            </Link>
+          </div>
+        </CardContent>
+      </Card>
+    </div>
+  );
+}
+
+export default Login;
